Refactor modifyGroup
